refactor arg_utils.py
