+++ python/llm/src/bigdl/llm/vllm/engine/async_llm_engine.py
@@ -37,7 +37,7 @@ from functools import partial
 from typing import (Any, Dict, Iterable, List, Optional, Set, Tuple, Type,
                     Union)
 from bigdl.llm.vllm.config import ModelConfig
-from bigdl.llm.vllm.utils.arg_utils import AsyncEngineArgs
+from bigdl.llm.vllm.engine.arg_utils import AsyncEngineArgs
 from bigdl.llm.vllm.engine.llm_engine import LLMEngine
 from bigdl.llm.vllm.logger import init_logger
 from bigdl.llm.vllm.structure.outputs import RequestOutput
